+++ src/conline.cpp
@@ -39,16 +39,26 @@ void conline::set_grid_objects()
 {
 
     opt_index.zeros(T + 1);
-    if(save_past_performance){
+    if (save_past_performance)
+    {
         past_performance.set_size(T);
-    }else{
+    }
+    else
+    {
         past_performance.set_size(1);
     }
     tmp_performance.zeros(D, P);
     cum_performance.zeros(X);
 
     predictions.zeros(T + T_E_Y, D, P);
-    predictions_tmp.set_size(T + T_E_Y);
+    if (save_predictions_grid)
+    {
+        predictions_grid.set_size(T + T_E_Y);
+    }
+    else
+    {
+        predictions_grid.set_size(1 + lead_time);
+    }
     weights_tmp.set_size(X);
     weights.set_size(T + 1);
 
@@ -115,12 +125,15 @@ void conline::set_grid_objects()
 
         weights(t).set_size(D, P, K);
 
-        if(save_past_performance){
+        if (save_past_performance)
+        {
             past_performance(t).set_size(D, P, X);
             past_performance(t).fill(datum::nan);
         }
-        
-        predictions_tmp(t).set_size(D, P, X);
+        predictions_grid(t).set_size(D, P, X);
+
+        // Store predictions w.r.t. grid for time t
+        cube tmp_preds_cube(D, P, X);
 
         for (unsigned int d = 0; d < D; d++)
         {
@@ -136,14 +149,14 @@ void conline::set_grid_objects()
                 mat weights_temp = weights_tmp(x).row(d);
 
                 // Forecasters prediction
-                vec predictions_temp = sum(weights_temp % experts_mat, 1);
+                vec tmp_preds_vec = sum(weights_temp % experts_mat, 1);
 
-                predictions_tmp(t)(span(d), span::all, span(x)) = predictions_temp;
+                tmp_preds_cube(span(d), span::all, span(x)) = tmp_preds_vec;
             }
-
-            // Final prediction
-            predictions.tube(t, d) = vectorise(predictions_tmp(t)(span(d), span::all, span(opt_index(t))));
         }
+        predictions_grid(t) = tmp_preds_cube;
+        // Final Prediction
+        predictions.row(t) = tmp_preds_cube.slice(opt_index(t));
     }
 
     start = lead_time;
@@ -154,16 +167,25 @@ void conline::learn()
 {
     Progress prog(T, trace);
     clock.tick("core");
+
+    for (unsigned int tp = 0; tp < predictions_grid.n_rows; tp++)
+    {
+        predictions_grid(tp).set_size(D, P, X);
+    }
+
     for (unsigned int t = start; t < T; t++)
     {
+
         weights(t).set_size(D, P, K);
 
-        if(save_past_performance)
+        if (save_past_performance)
             past_performance(t).set_size(D, P, X);
-        
-        predictions_tmp(t).set_size(D, P, X);
 
         clock.tick("loss");
+
+        // Store predictions w.r.t. grid for time t
+        cube tmp_preds_cube(D, P, X);
+
         for (unsigned int d = 0; d < D; d++)
         {
             // Save final weights weights_tmp
@@ -177,18 +199,40 @@ void conline::learn()
             {
 
                 mat weights_temp = weights_tmp(x).row(d);
-                vec predictions_temp = sum(weights_temp % experts_mat, 1);
+                vec tmp_preds_vec = sum(weights_temp % experts_mat, 1);
 
                 // Sort predictions if quantile_crossing is prohibited
                 if (!allow_quantile_crossing)
                 {
-                    predictions_temp = arma::sort(predictions_temp, "ascend", 0);
+                    tmp_preds_vec = arma::sort(tmp_preds_vec, "ascend", 0);
                 }
-                predictions_tmp(t)(span(d), span::all, span(x)) = predictions_temp;
+                tmp_preds_cube(span(d), span::all, span(x)) = tmp_preds_vec;
             }
-            // Forecasters prediction
-            predictions.tube(t, d) = vectorise(predictions_tmp(t)(span(d), span::all, span(opt_index(t))));
         }
+
+        // Define which observation to base the weight update on. That is the
+        // most recent prediction unless lead_time is greater than 0.
+        arma::uword predictions_tmp_idx;
+        if (save_predictions_grid)
+        {
+            predictions_tmp_idx = t - lead_time;
+            predictions_grid(t) = tmp_preds_cube;
+        }
+        else
+        {
+            // If save_predictions_grid is false, the first observation is
+            // allways the one we base the weight update on. This is because
+            // the size of predictions_grid is only 1 + lead_time.
+            predictions_tmp_idx = 0;
+            for (unsigned int tp = 0; tp < predictions_grid.n_rows - 1; tp++)
+            {
+                predictions_grid(tp) = predictions_grid(tp + 1);
+            }
+            predictions_grid(predictions_grid.n_rows - 1) = tmp_preds_cube;
+        }
+
+        // Final prediction
+        predictions.row(t) = tmp_preds_cube.slice(opt_index(t));
 
         clock.tock("loss");
         for (unsigned int x = 0; x < X; x++)
@@ -208,24 +252,24 @@ void conline::learn()
 #pragma omp parallel for
                 for (unsigned int p = 0; p < P; p++)
                 {
-                    tmp_performance(d,p) = loss(y(t, d),
-                                                        predictions_tmp(t - lead_time)(d, p, x),
-                                                        9999,           // where evaluate loss_gradient
-                                                        loss_function,  // method
-                                                        tau(p),         // tau
-                                                        loss_parameter, // alpha
-                                                        false);
-                    
+                    tmp_performance(d, p) = loss(y(t, d),
+                                                 predictions_grid(predictions_tmp_idx)(d, p, x),
+                                                 9999,           // where evaluate loss_gradient
+                                                 loss_function,  // method
+                                                 tau(p),         // tau
+                                                 loss_parameter, // alpha
+                                                 false);
+
                     if (params["loss_share"](x) != 1)
                     {
                         for (unsigned int k = 0; k < K; k++)
                         {
                             lexp_int(p, k) = loss(y(t, d),
                                                   experts(t)(d, p, k),
-                                                  predictions_tmp(t - lead_time)(d, p, x), // where evaluate loss_gradient
-                                                  loss_function,                           // method
-                                                  tau(p),                                  // tau
-                                                  loss_parameter,                          // alpha
+                                                  predictions_grid(predictions_tmp_idx)(d, p, x), // where evaluate loss_gradient
+                                                  loss_function,                                  // method
+                                                  tau(p),                                         // tau
+                                                  loss_parameter,                                 // alpha
                                                   loss_gradient);
                         }
 
@@ -245,11 +289,11 @@ void conline::learn()
                         lexp.row(p) = lexp_ext.row(p);
                     }
                     lfor(p) = loss(y(t, d),
-                                   predictions_tmp(t - lead_time)(d, p, x),
-                                   predictions_tmp(t - lead_time)(d, p, x), // where to evaluate loss_gradient
-                                   loss_function,                           // method
-                                   tau(p),                                  // tau
-                                   loss_parameter,                          // alpha
+                                   predictions_grid(predictions_tmp_idx)(d, p, x),
+                                   predictions_grid(predictions_tmp_idx)(d, p, x), // where to evaluate loss_gradient
+                                   loss_function,                                  // method
+                                   tau(p),                                         // tau
+                                   loss_parameter,                                 // alpha
                                    loss_gradient);
                 }
 
@@ -400,7 +444,7 @@ void conline::learn()
             }
 
 #pragma omp parallel for
-            // Enshure that constraints hold
+            //  Enshure that constraints hold
             for (unsigned int p = 0; p < P; p++)
             {
                 for (unsigned int d = 0; d < D; d++)
@@ -414,15 +458,15 @@ void conline::learn()
                         accu(weights_tmp(x)(span(d), span(p), span::all));
                 }
             }
-            if(save_past_performance)
-                past_performance(t).slice(x) =  tmp_performance;        
+            if (save_past_performance)
+                past_performance(t).slice(x) = tmp_performance;
             R_CheckUserInterrupt();
 
-        // Apply forget
-        cum_performance(x) *= (1 - forget_past_performance);
-        // Add new loss
-        cum_performance(x) += accu(tmp_performance) / double(D*P);
-        
+            // Apply forget
+            cum_performance(x) *= (1 - forget_past_performance);
+            // Add new loss
+            cum_performance(x) += accu(tmp_performance) / double(D * P);
+
         } // x
 
         opt_index(t + 1) = cum_performance.index_min();
@@ -440,14 +484,14 @@ void conline::learn()
         {
             mat experts_mat = experts(t).row(d);
             mat weights_temp = weights(T).row(d);
-            vec predictions_temp = sum(weights_temp % experts_mat, 1);
+            vec tmp_preds_vec = sum(weights_temp % experts_mat, 1);
 
             // Sort predictions if quantile_crossing is prohibited
             if (!allow_quantile_crossing)
             {
-                predictions_temp = arma::sort(predictions_temp, "ascend", 0);
+                tmp_preds_vec = arma::sort(tmp_preds_vec, "ascend", 0);
             }
-            predictions.tube(t, d) = predictions_temp;
+            predictions.tube(t, d) = tmp_preds_vec;
         }
     }
 
@@ -491,10 +535,8 @@ Rcpp::List conline::output()
 {
     clock.tick("wrangle");
 
-    // // 1-Indexing for R-Output
+    // 1-Indexing for R-Output
     opt_index += 1;
-
-    // Rcpp::NumericMatrix parametergrid_out = Rcpp::wrap(param_grid);
 
     Rcpp::List model_data = Rcpp::List::create(
         Rcpp::Named("y") = y,
@@ -508,11 +550,13 @@ Rcpp::List conline::output()
         Rcpp::Named("loss_gradient") = loss_gradient,
         Rcpp::Named("method") = method,
         Rcpp::Named("forget_past_performance") = forget_past_performance,
-        Rcpp::Named("allow_quantile_crossing") = allow_quantile_crossing);
+        Rcpp::Named("allow_quantile_crossing") = allow_quantile_crossing,
+        Rcpp::Named("save_past_performance") = save_past_performance,
+        Rcpp::Named("save_predictions_grid") = save_predictions_grid);
 
     Rcpp::List model_objects = Rcpp::List::create(
         Rcpp::Named("weights_tmp") = weights_tmp,
-        Rcpp::Named("predictions_tmp") = predictions_tmp,
+        Rcpp::Named("predictions_grid") = predictions_grid,
         Rcpp::Named("cum_performance") = cum_performance,
         Rcpp::Named("hat_pr") = hat_pr,
         Rcpp::Named("hat_mv") = hat_mv,
@@ -599,22 +643,11 @@ void conline::init_update(
     opt_index -= 1;
     opt_index.resize(T + 1);
 
-    if(save_past_performance){
-        past_performance.set_size(T);
-        past_performance.rows(0, start - 1) =
-            Rcpp::as<arma::field<cube>>(object["past_performance"]);
-    }else{
-        past_performance.set_size(1);
-    }
-
-    tmp_performance.zeros(D,P);
+    tmp_performance.zeros(D, P);
     cum_performance = Rcpp::as<arma::vec>(model_objects["cum_performance"]);
 
     weights_tmp =
         Rcpp::as<arma::field<cube>>(model_objects["weights_tmp"]);
-
-    predictions_tmp.set_size(T + T_E_Y);
-    predictions_tmp.rows(0, old_experts.n_rows - 1) = Rcpp::as<arma::field<cube>>(model_objects["predictions_tmp"]);
 
     // // Output Objects
     predictions = Rcpp::as<arma::cube>(object["predictions"]);
@@ -645,6 +678,31 @@ void conline::init_update(
 
     forget_past_performance = model_parameters["forget_past_performance"];
     allow_quantile_crossing = model_parameters["allow_quantile_crossing"];
+
+    save_past_performance = model_parameters["save_past_performance"];
+    save_predictions_grid = model_parameters["save_predictions_grid"];
+
+    if (save_past_performance)
+    {
+        past_performance.set_size(T);
+        past_performance.rows(0, start - 1) =
+            Rcpp::as<arma::field<cube>>(object["past_performance"]);
+    }
+    else
+    {
+        past_performance.set_size(1);
+    }
+
+    if (save_predictions_grid)
+    {
+        predictions_grid.set_size(T + T_E_Y);
+        predictions_grid.rows(0, old_experts.n_rows - 1) = Rcpp::as<arma::field<cube>>(model_objects["predictions_grid"]);
+    }
+    else
+    {
+        predictions_grid.set_size(1 + lead_time);
+        predictions_grid = Rcpp::as<arma::field<cube>>(model_objects["predictions_grid"]);
+    }
 
     loss_for.zeros(T, D, P);
     loss_for.rows(0, start - 1) =